refactored all tests, improved documentation Signed-off-by: Rod Anami <rod.anami@kyndryl.com>
